Cast using unsigned instead of uint16_t.
There's no difference in behavior but we don't care about the size we're
extending the type to.

+++ src/metrics/include/metrics/operation.hpp
@@ -62,9 +62,9 @@ struct OperationEvent {
         out << ",size:" << event.size;
         out << ",errors:" << event.errors;
         out << ",duration:" << event.duration;
-        // Casting to uint8_t causes ostream to use its unsigned char overload and treat the value
-        // as invisible text.
-        out << ",outcome:" << static_cast<uint16_t>(event.outcome);
+        // Casting to uint8_t directly causes ostream to use its unsigned char overload and treat
+        // the value as invisible text.
+        out << ",outcome:" << static_cast<unsigned>(event.outcome);
         out << "}";
         return out;
     }
